Refacto AASM Project in a Concern.
